+++ AutoClicker/resource.h
@@ -2,3 +2,4 @@
 
 #define IDI_ICON1                       101
 #define IDR_MC_DLL                      102
+#define IDR_GLFW_PROXY                  103